+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- In-App Review test in DebugHomePanel [#2131](https://github.com/rokwire/illinois-app/issues/2131).
+
 ## [4.1.5] - 2022-07-28
 ### Added
 - Display types "All", "Multiple" and "Single" for events [#2124](https://github.com/rokwire/illinois-app/issues/2124).